d2cli: Move main code into d2cli so it can be imported from tests
Updates #903
