+++ d2cli/static/watch.js
